Move auxiliary functions into utils package.
