+++ simulation.py
@@ -2,7 +2,7 @@
 
 from simulations.simslack import run_sim
 from argparse import ArgumentParser, FileType
-from utils import calculate_k, get_rts
+from utils.rts import calculate_k, get_rts
 from schedtests import josephp
 import math
 import sys
@@ -33,8 +33,9 @@ def run_single_simulation(rts, args):
     sim_result = run_sim(params, callback=None, sink=True, retrieve_model=True)
 
     if sim_result["error"]:
-        print("Error: {0}".format(sim_result["error_msg"]))
-        sys.exit(1)
+        print("Error: RTS {0}, {1}".format(rts["id"], sim_result["error_msg"]), file=sys.stderr)
+        if args.exit_on_error:
+            sys.exit(1)
 
 
 def get_args():
@@ -44,6 +45,7 @@ def get_args():
     parser.add_argument("--scheduler", nargs=1, type=str, help="Scheduling algorithm")
     parser.add_argument("--instance-count", type=int, help="Number of task instances to simulate.")
     parser.add_argument("--ss-methods", nargs='+', type=str, help="Slack Stealing methods.")
+    parser.add_argument("--exit-on-error", default=False, action="store_true", help="Exit if simulation error.")
     return parser.parse_args()
 
 
@@ -54,8 +56,11 @@ def main():
 
     args = get_args()
 
-    for rts in get_rts(args.file):
-        run_single_simulation(rts, args)
+    try:
+        for rts in get_rts(args.file):
+            run_single_simulation(rts, args)
+    except KeyboardInterrupt:
+        sys.exit(1)
 
 
 if __name__ == '__main__':
